seeding from user-provided coordinates

+++ cpp/core/dwi/tractography/seeding/base.h
@@ -62,7 +62,7 @@ public:
   virtual ~Base() {}
 
   default_type vol() const { return volume; }
-  uint32_t num() const { return count; }
+  size_t num() const { return count; }
   bool is_finite() const { return count; }
   std::string get_type() const { return type; }
   std::string get_name() const { return name; }
@@ -79,7 +79,7 @@ public:
 protected:
   // Finite seeds are defined by the number of seeds; non-limited are defined by volume
   float volume;
-  uint32_t count;
+  size_t count;
   mutable std::mutex mutex;
   const std::string type; // Text describing the type of seed this is
